requirements.txt
# =============================================================================
# AI Metadata Enricher - Orchestrator Dependencies
# =============================================================================
# Minimal dependencies for the structural orchestrator (no LLM, no RAG).

# Azure Service Bus SDK - peek-lock message consumption
azure-servicebus>=7.12.0,<8.0.0

# Azure Identity - Managed Identity authentication
azure-identity>=1.15.0,<2.0.0

# Azure Monitor OpenTelemetry - Application Insights integration
azure-monitor-opentelemetry>=1.3.0,<2.0.0
